Ensure -framework Carbon go into gthread pkg-config file.
When a user links against the gthread library they need to
also link against the Carbon framework. We fix this up by
passing it through from configure to the .pc file.

This is a bit of an expediency as configure itself is generated,
so the real fix should probably go in to configure.ac